requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.2
drf-extra-fields==3.0.3
djangorestframework-gis==0.16
Markdown==3.3.3
